Move credentials encoding into one service
