+++ src/app/Shared/Services/Login.service.tsx
@@ -36,9 +36,10 @@
  * SOFTWARE.
  */
 import { Base64 } from 'js-base64';
-import { Observable, ObservableInput, of, ReplaySubject } from 'rxjs';
+import { combineLatest, Observable, ObservableInput, of, ReplaySubject } from 'rxjs';
 import { fromFetch } from 'rxjs/fetch';
 import { catchError, concatMap, first, map, tap } from 'rxjs/operators';
+import { TargetService } from './Target.service';
 
 export enum SessionState {
   NO_USER_SESSION,
@@ -57,7 +58,7 @@ export class LoginService {
   private readonly sessionState = new ReplaySubject<SessionState>(1);
   readonly authority: string;
 
-  constructor() {
+  constructor(private readonly target: TargetService) {
     let apiAuthority = process.env.CRYOSTAT_AUTHORITY;
     if (!apiAuthority) {
       apiAuthority = '';
@@ -76,11 +77,7 @@ export class LoginService {
   }
 
   checkAuth(token: string, method: string, rememberMe = false): Observable<boolean> {
-
-    if (method === 'Basic') {
-      token = Base64.encodeURL(token);
-    }
-
+    token = Base64.encodeURL(token);
     token = this.useCacheItemIfAvailable(this.TOKEN_KEY, token);
 
     return fromFetch(`${this.authority}/api/v2.1/auth`, {
@@ -117,17 +114,39 @@ export class LoginService {
     );
   }
 
-  getToken(): Observable<string> {
-    return this.token.asObservable();
-  }
-
   getAuthHeaders(token: string, method: string): Headers {
     const headers = new window.Headers();
     if (!!token && !!method) {
       headers.set('Authorization', `${method} ${token}`)
     }
-
     return headers;
+  }
+
+  getHeaders(): Observable<Headers> {
+    const authorization = combineLatest([this.getToken(), this.getAuthMethod()])
+    .pipe(
+      map((parts: [string, string]) => this.getAuthHeaders(parts[0], parts[1])),
+      first(),
+    );
+    return combineLatest([authorization, this.target.target()])
+    .pipe(
+      first(),
+      map(parts => {
+        const headers = parts[0];
+        const target = parts[1];
+        if (!!target && !!target.connectUrl && this.target.hasCredentials(target.connectUrl)) {
+          const credentials = this.target.getCredentials(target.connectUrl);
+          if (credentials) {
+            headers.set('X-JMX-Authorization', `Basic ${Base64.encodeURL(credentials)}`);
+          }
+        }
+        return headers;
+      })
+    );
+  }
+
+  getToken(): Observable<string> {
+    return this.token.asObservable();
   }
 
   getAuthMethod(): Observable<string> {
